requirements-base.txt
asyncpg
# Kafka
confluent-kafka>=2.0.0
aiokafka>=0.8.0
TA-Lib
# Database
psycopg2-binary>=2.9.5
SQLAlchemy>=2.0.0

# Data processing
pandas>=1.5.3
numpy>=1.23.5

# Market data
yfinance>=0.2.18
twelvedata>=1.2.11

# API & Web
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
websockets>=11.0.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.28.2
tenacity>=8.2.2
aiohttp>=3.8.4

# Logging & monitoring
structlog>=23.1.0